+++ ing-software-dunamys-v3/ing-software-dunamys-v3/src/main/java/com/uia/ing/soft/olda/dunamys/ing_software_dunamys/Controller/EstadoFacturaControlador.java
@@ -35,7 +35,7 @@ public class EstadoFacturaControlador {
     }
     @DeleteMapping("/{userId}/{id}")
     public ResponseEntity<String> eliminarEstadoFactura(@PathVariable Integer userId
-                                                      , @PathVariable Long id){
+                                                      , @PathVariable Integer id){
         Optional<EstadoFactura> estadoFactura = servicio.findById(id);
         if (estadoFactura.isPresent()) {
             servicio.delete(id);
