+++ ing-software-dunamys-v3/ing-software-dunamys-v3/src/main/java/com/uia/ing/soft/olda/dunamys/ing_software_dunamys/Repository/DetalleFacturaRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.uia.ing.soft.olda.dunamys.ing_software_dunamys.Model.DetalleFactura;
 
 @Repository
-public interface DetalleFacturaRepositorio extends JpaRepository<DetalleFactura, Long>{
+public interface DetalleFacturaRepositorio extends JpaRepository<DetalleFactura, Integer>{
 
 }
