+++ ing-software-dunamys-v3/ing-software-dunamys-v3/src/main/java/com/uia/ing/soft/olda/dunamys/ing_software_dunamys/Repository/EstadoReservacionRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.uia.ing.soft.olda.dunamys.ing_software_dunamys.Model.EstadoReservacion;
 
 @Repository
-public interface EstadoReservacionRepositorio extends JpaRepository<EstadoReservacion, Long>{
+public interface EstadoReservacionRepositorio extends JpaRepository<EstadoReservacion, Integer>{
 
 }
