+++ ing-software-dunamys-v3/ing-software-dunamys-v3/src/main/java/com/uia/ing/soft/olda/dunamys/ing_software_dunamys/Repository/PersonaRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.uia.ing.soft.olda.dunamys.ing_software_dunamys.Model.Persona;
 
 @Repository
-public interface PersonaRepositorio extends JpaRepository<Persona, Long>{
+public interface PersonaRepositorio extends JpaRepository<Persona, Integer>{
 
 }
