+++ ing-software-dunamys-v3/ing-software-dunamys-v3/src/main/java/com/uia/ing/soft/olda/dunamys/ing_software_dunamys/Repository/TipoHabitacionRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.uia.ing.soft.olda.dunamys.ing_software_dunamys.Model.TipoHabitacion;
 
 @Repository
-public interface TipoHabitacionRepositorio extends JpaRepository<TipoHabitacion, Long>{
+public interface TipoHabitacionRepositorio extends JpaRepository<TipoHabitacion, Integer>{
 
 }
