+++ ing-software-dunamys-v3/ing-software-dunamys-v3/src/main/java/com/uia/ing/soft/olda/dunamys/ing_software_dunamys/Repository/TipoReservacionRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.uia.ing.soft.olda.dunamys.ing_software_dunamys.Model.TipoReservacion;
 
 @Repository
-public interface TipoReservacionRepositorio extends JpaRepository<TipoReservacion, Long>{
+public interface TipoReservacionRepositorio extends JpaRepository<TipoReservacion, Integer>{
 
 }
