+++ ing-software-dunamys-v3/ing-software-dunamys-v3/src/main/java/com/uia/ing/soft/olda/dunamys/ing_software_dunamys/Service/EstadoFacturaServicio.java
@@ -7,7 +7,7 @@ import com.uia.ing.soft.olda.dunamys.ing_software_dunamys.Model.EstadoFactura;
 import com.uia.ing.soft.olda.dunamys.ing_software_dunamys.Repository.EstadoFacturaRepositorio;
 
 @Service
-public class EstadoFacturaServicio extends BaseService<EstadoFactura, Long>{
+public class EstadoFacturaServicio extends BaseService<EstadoFactura, Integer>{
 
     @Autowired
     private EstadoFacturaRepositorio repo;
